+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added tag classes for group D.2
 -   Added tag classes for group D.3
 -   Added tag classes for group E
+-   Added tag classes for group E.1
+-   Added tag classes for group E.1.1
 -   Base classes of the Builder pattern are added for the generation of Factura Electronica
 -   Factura Electronica generation test added